Setup robot part for communication and refactor webserver
