+++ server/aimbot/web/web.py
@@ -1,0 +1,8 @@
+from flask import Blueprint, render_template
+
+web = Blueprint('web', __name__, template_folder="../../templates")
+
+
+@web.route('/')
+def start():
+    return render_template('index.html')
