+++ server/requirements.txt
@@ -3,4 +3,5 @@ Flask==2.0.1
 itsdangerous==2.0.1
 Jinja2==3.0.1
 MarkupSafe==2.0.1
+smbus2==0.4.1
 Werkzeug==2.0.1
